solved 3rd lesson
The "debugger;" directive which defines a breakpoint at the first line of the code only works if the command used to run the program is similar to "node-debug main.js file".

To be able to run the program as said it is necessary to install the inspector with "npm install -g node-inspector".

When the breakpoint is acheived in the execution thread a nem window will be opened in the default browser and a console similar to the chrome javascript
debugger console will be presented in the page.